+++ footer.php
@@ -1,7 +1,6 @@
-    <?php get_template_part('components/layout/footer/footer'); ?>
-    </div><!-- #page -->
-
-    <?php wp_footer(); ?>
+<?php
+get_template_part('components/site-footer/site-footer');
+wp_footer(); ?>
 
 </body>
 </html>